+++ src/objects/DepositedCharge.hpp
@@ -13,7 +13,7 @@ namespace allpix {
     // object definition
     class DepositedCharge : public SensorCharge {
     public:
-        DepositedCharge(ROOT::Math::XYZPoint position, unsigned int charge, long double event_time)
+        DepositedCharge(ROOT::Math::XYZPoint position, unsigned int charge, double event_time)
             : SensorCharge(std::move(position), charge, event_time) {}
 
         ClassDef(DepositedCharge, 1);
